Split sync and non sync utils
