+++ tests/CMakeLists.txt
@@ -5,6 +5,7 @@ set(HEADERS
     util/index_helpers.hpp
     util/templated_test_case.hpp
     util/test_file.hpp
+    util/test_utils.hpp
 )
 
 set(SOURCES
@@ -25,6 +26,7 @@ set(SOURCES
 
     util/event_loop.cpp
     util/test_file.cpp
+    util/test_utils.cpp
 )
