+++ tests/sync/file.cpp
@@ -20,6 +20,7 @@
 
 #include "shared_realm.hpp"
 #include "sync/sync_manager.hpp"
+#include "util/test_utils.hpp"
 #include <realm/util/file.hpp>
 #include <realm/util/scope_exit.hpp>
