+++ tests/sync/sync_test_utils.hpp
@@ -26,38 +26,9 @@
 
 namespace realm {
 
-/// Open a Realm at a given path, creating its files.
-bool create_dummy_realm(std::string path);
-void reset_test_directory(const std::string& base_path);
 bool results_contains_user(SyncUserMetadataResults& results, const std::string& identity, const std::string& auth_server);
 bool results_contains_original_name(SyncFileActionMetadataResults& results, const std::string& original_name);
-std::string tmp_dir();
-std::vector<char> make_test_encryption_key(const char start = 0);
 
 } // namespace realm
-
-#define REQUIRE_DIR_EXISTS(macro_path) do { \
-    DIR *dir_listing = opendir((macro_path).c_str()); \
-    CHECK(dir_listing); \
-    if (dir_listing) closedir(dir_listing); \
-} while (0)
-
-#define REQUIRE_DIR_DOES_NOT_EXIST(macro_path) do { \
-    DIR *dir_listing = opendir((macro_path).c_str()); \
-    CHECK(dir_listing == NULL); \
-    if (dir_listing) closedir(dir_listing); \
-} while (0)
-
-#define REQUIRE_REALM_EXISTS(macro_path) do { \
-	REQUIRE(realm::util::File::exists(macro_path)); \
-	REQUIRE(realm::util::File::exists((macro_path) + ".lock")); \
-	REQUIRE_DIR_EXISTS((macro_path) + ".management"); \
-} while (0)
-
-#define REQUIRE_REALM_DOES_NOT_EXIST(macro_path) do { \
-	REQUIRE(!realm::util::File::exists(macro_path)); \
-	REQUIRE(!realm::util::File::exists((macro_path) + ".lock")); \
-	REQUIRE_DIR_DOES_NOT_EXIST((macro_path) + ".management"); \
-} while (0)
 
 #endif // REALM_SYNC_TEST_UTILS_HPP
